+++ app/Views/accesscontrol.php
@@ -1,47 +1,53 @@
 <?= $this->extend('layout') ?>
 <?= $this->section('main') ?>
 <?php if ($ismobile === false) { ?>
-    <div class="uk-child-width-auto uk-flex-between uk-flex-middle" uk-grid>
-        <div>
-            <h1>Hak Akses</h1>
-        </div>
-        <div>
-            <a class="uk-button uk-button-large uk-button-secondary" href="#modaladd" uk-toggle>Tambah Hak Akses</a>
-        </div>
-
-        <div id="modaladd" uk-modal>
-            <div class="uk-modal-dialog uk-margin-auto-vertical" uk-overflow-auto>
-                <button class="uk-modal-close-default" type="button" uk-close></button>
-                <div class="uk-modal-header">
-                    <h2 class="uk-modal-title">Add Permissions</h2>
-                </div>
-                <form method="post" action="users/create/access-control">
-                    <div class="uk-modal-body">
-                        <div class="uk-margin">
-                            <input class="uk-input" type="text" name="group" placeholder="Groups" aria-label="Input">
-                        </div>
-
-                        <div class="uk-margin">
-                            <input class="uk-input" type="text" name="description" placeholder="Description" aria-label="Input">
-                        </div>
-
-                        <div class="uk-margin uk-grid-small uk-child-width-auto uk-grid">
-                            <?php foreach ($permissions as $permission) { ?>
-                                <label><input class="uk-checkbox" name="permission[<?= $permission['id'] ?>]" value="<?= $permission['id'] ?>" type="checkbox"> <?= $permission['name'] ?></label>
-                            <?php } ?>
-                        </div>
-                        <div class="uk-modal-footer uk-text-right">
-                            <button class="uk-button uk-button-default uk-modal-close" type="button">Cancel</button>
-                            <button class="uk-button uk-button-primary" type="submit">Save</button>
-                        </div>
-                    </div>
-                </form>
+    <?php if ($authorize->hasPermission('admin.user.read', $uid)) { ?>
+        <div class="uk-child-width-auto uk-flex-between uk-flex-middle" uk-grid>
+            <div>
+                <h1>Hak Akses</h1>
             </div>
+            <?php if ($authorize->hasPermission('admin.user.create', $uid)) { ?>
+                <div>
+                    <a class="uk-button uk-button-large uk-button-secondary" href="#modaladd" uk-toggle>Tambah Hak Akses</a>
+                </div>
+            <?php } ?>
+
+            <?php if ($authorize->hasPermission('admin.user.create', $uid)) { ?>
+                <div id="modaladd" uk-modal>
+                    <div class="uk-modal-dialog uk-margin-auto-vertical" uk-overflow-auto>
+                        <button class="uk-modal-close-default" type="button" uk-close></button>
+                        <div class="uk-modal-header">
+                            <h2 class="uk-modal-title">Add Permissions</h2>
+                        </div>
+                        <form method="post" action="users/create/access-control">
+                            <div class="uk-modal-body">
+                                <div class="uk-margin">
+                                    <input class="uk-input" type="text" name="group" placeholder="Groups" aria-label="Input">
+                                </div>
+
+                                <div class="uk-margin">
+                                    <input class="uk-input" type="text" name="description" placeholder="Description" aria-label="Input">
+                                </div>
+
+                                <div class="uk-margin uk-grid-small uk-child-width-auto uk-grid">
+                                    <?php foreach ($permissions as $permission) { ?>
+                                        <label><input class="uk-checkbox" name="permission[<?= $permission['id'] ?>]" value="<?= $permission['id'] ?>" type="checkbox"> <?= $permission['name'] ?></label>
+                                    <?php } ?>
+                                </div>
+                                <div class="uk-modal-footer uk-text-right">
+                                    <button class="uk-button uk-button-default uk-modal-close" type="button">Cancel</button>
+                                    <button class="uk-button uk-button-primary" type="submit">Save</button>
+                                </div>
+                            </div>
+                        </form>
+                    </div>
+                </div>
+            <?php } ?>
+
         </div>
 
-    </div>
-
-    <?= view('Views/Auth/_message_block') ?>
+        <?= view('Views/Auth/_message_block') ?>
+    <?php } ?>
 <?php } else { ?>
     <h1 class="uk-text-center">Hak Akses</h1>
     <div class="uk-margin uk-text-center"><button class="uk-button uk-button-large uk-button-secondary">Tambah Hak Akses</button></div>
@@ -69,8 +75,12 @@
                 if (!in_array($group->name, $grouparr)) { ?>
                     <div class="uk-card-footer">
                         <div class="uk-child-width-auto" uk-grid>
-                            <div><a class="uk-button uk-button-primary" uk-icon="pencil" href="#modalupdate<?= $group->id; ?>" uk-toggle></a></div>
-                            <div><a class="uk-button uk-button-danger" href="users/delete/access-control/<?= $group->id; ?>" onClick="return confirm('<?= lang('Global.deleteConfirm') ?>')" uk-icon="trash"></a></div>
+                            <?php if ($authorize->hasPermission('admin.user.edit', $uid)) { ?>
+                                <div><a class="uk-button uk-button-primary" uk-icon="pencil" href="#modalupdate<?= $group->id; ?>" uk-toggle></a></div>
+                            <?php } ?>
+                            <?php if ($authorize->hasPermission('admin.user.delete', $uid)) { ?>
+                                <div><a class="uk-button uk-button-danger" href="users/delete/access-control/<?= $group->id; ?>" onClick="return confirm('<?= lang('Global.deleteConfirm') ?>')" uk-icon="trash"></a></div>
+                            <?php } ?>
                         </div>
                     </div>
                 <?php } ?>
@@ -78,56 +88,58 @@
         </div>
 
         <!-- update permission modal -->
-        <div id="modalupdate<?= $group->id; ?>" uk-modal>
-            <div class="uk-modal-dialog uk-margin-auto-vertical" uk-overflow-auto>
-                <button class="uk-modal-close-default" type="button" uk-close></button>
+        <?php if ($authorize->hasPermission('admin.user.edit', $uid)) { ?>
+            <div id="modalupdate<?= $group->id; ?>" uk-modal>
+                <div class="uk-modal-dialog uk-margin-auto-vertical" uk-overflow-auto>
+                    <button class="uk-modal-close-default" type="button" uk-close></button>
 
-                <div class="uk-modal-header">
-                    <h2 class="uk-modal-title">Update Permissions</h2>
-                </div>
+                    <div class="uk-modal-header">
+                        <h2 class="uk-modal-title">Update Permissions</h2>
+                    </div>
 
-                <div class="uk-modal-body">
-                    <form method="post" action="users/update/access/<?= $group->id; ?>">
+                    <div class="uk-modal-body">
+                        <form method="post" action="users/update/access/<?= $group->id; ?>">
 
-                        <div class="uk-margin">
-                            <input class="uk-input" type="text" name="name" value="<?= $group->name; ?>" aria-label="Input">
-                            <input class="uk-input" type="number" name="groupid" value="<?= $group->id; ?>" aria-label="Input" hidden>
-                        </div>
+                            <div class="uk-margin">
+                                <input class="uk-input" type="text" name="name" value="<?= $group->name; ?>" aria-label="Input">
+                                <input class="uk-input" type="number" name="groupid" value="<?= $group->id; ?>" aria-label="Input" hidden>
+                            </div>
 
-                        <div class="uk-margin">
-                            <input class="uk-input" type="text" name="description" value="<?= $group->description; ?>" aria-label="Input">
-                        </div>
+                            <div class="uk-margin">
+                                <input class="uk-input" type="text" name="description" value="<?= $group->description; ?>" aria-label="Input">
+                            </div>
 
-                        <div class="uk-margin">
-                            <?php
-                            foreach ($permissions as $permission) {
-                                $permarr = [];
-                                $grouppermissions = $GroupModel->getPermissionsForGroup($group->id);
-                                foreach ($grouppermissions as $grouppermission) {
-                                    $permarr[] = $grouppermission->id;
+                            <div class="uk-margin">
+                                <?php
+                                foreach ($permissions as $permission) {
+                                    $permarr = [];
+                                    $grouppermissions = $GroupModel->getPermissionsForGroup($group->id);
+                                    foreach ($grouppermissions as $grouppermission) {
+                                        $permarr[] = $grouppermission->id;
+                                    }
+                                    if (in_array($permission['id'], $permarr)) {
+                                        $checked = 'checked';
+                                    } else {
+                                        $checked = '';
+                                    }
+                                ?>
+                                    <div>
+                                        <label><input class="uk-checkbox" name="permission[]" value="<?= $permission['id'] ?>" type="checkbox" <?= $checked ?>> <?= $permission['description'] ?></label>
+                                    </div>
+                                <?php
                                 }
-                                if (in_array($permission['id'], $permarr)) {
-                                    $checked = 'checked';
-                                } else {
-                                    $checked = '';
-                                }
-                            ?>
-                                <div>
-                                    <label><input class="uk-checkbox" name="permission[]" value="<?= $permission['id'] ?>" type="checkbox" <?= $checked ?>> <?= $permission['description'] ?></label>
-                                </div>
-                            <?php
-                            }
-                            ?>
-                        </div>
+                                ?>
+                            </div>
 
-                        <div class="uk-modal-footer uk-text-right">
-                            <button class="uk-button uk-button-default uk-modal-close" type="button">Cancel</button>
-                            <button class="uk-button uk-button-primary" type="submit">Save</button>
-                        </div>
-                    </form>
+                            <div class="uk-modal-footer uk-text-right">
+                                <button class="uk-button uk-button-default uk-modal-close" type="button">Cancel</button>
+                                <button class="uk-button uk-button-primary" type="submit">Save</button>
+                            </div>
+                        </form>
+                    </div>
                 </div>
             </div>
-        </div>
+        <?php } ?>
         <!-- end update permission modal  -->
 
     <?php } ?>
